update docker containers
