+++ docker/modules/Dockerfile.faux1
@@ -39,7 +39,7 @@ RUN $AG update && $AG install isc-dhcp-client ethtool network-manager netcat cur
 # Additional OS dependencies
 RUN $AG update && $AG install -y telnetd && $AG install xinetd nginx
 
-COPY subset/ntp/NTPClient NTPClient
+COPY subset/network/NTPClient NTPClient
 RUN cd NTPClient && ./gradlew build
 COPY subset/network/TransportClient TransportClient
