+++ subset/network/Dockerfile.test_network
@@ -1,8 +1,14 @@
 FROM daqf/aardvark:latest
 
-RUN $AG update && $AG install python netcat
+RUN $AG update && $AG install python python-setuptools python-pip netcat
+
+RUN pip install scapy
+
+
+COPY subset/ntp/ntp_tests.py .
 
 COPY subset/network/network_tests.py .
+COPY subset/network/ntp_tests.py .
 COPY subset/network/test_network .
 
 CMD ["./test_network"]
