--- subset/ntp/NTPClient/build.gradle
@@ -1,22 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group 'ntp_client'
-version '1.0-SNAPSHOT'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.13'
-}
-
-jar {
-    manifest {
-        attributes 'Main-Class': 'Main'
-    }
-}
